+++ addons/mail/models/mail_template.py
@@ -80,6 +80,7 @@ except ImportError:
 class MailTemplate(models.Model):
     "Templates for sending email"
     _name = "mail.template"
+    _inherit = ['mail.render.mixin']
     _description = 'Email Templates'
     _order = 'name'
 
@@ -134,20 +135,6 @@ class MailTemplate(models.Model):
                                            "emails created from this template")
     auto_delete = fields.Boolean('Auto Delete', default=True, help="Permanently delete this email after sending it, to save space")
 
-    # Fake fields used to implement the placeholder assistant
-    model_object_field = fields.Many2one('ir.model.fields', string="Field",
-                                         help="Select target field from the related document model.\n"
-                                              "If it is a relationship field you will be able to select "
-                                              "a target field at the destination of the relationship.")
-    sub_object = fields.Many2one('ir.model', 'Sub-model', readonly=True,
-                                 help="When a relationship field is selected as first field, "
-                                      "this field shows the document model the relationship goes to.")
-    sub_model_object_field = fields.Many2one('ir.model.fields', 'Sub-field',
-                                             help="When a relationship field is selected as first field, "
-                                                  "this field lets you select the target field within the "
-                                                  "destination document model (sub-model).")
-    null_value = fields.Char('Default Value', help="Optional value to use if the target field is empty")
-    copyvalue = fields.Char('Placeholder Expression', help="Final placeholder expression, to be copy-pasted in the desired template field.")
     scheduled_date = fields.Char('Scheduled Date', help="If set, the queue manager will send the email after the date. If not set, the email will be send as soon as possible. Jinja2 placeholders may be used.")
 
     @api.onchange('model_id')
@@ -157,42 +144,6 @@ class MailTemplate(models.Model):
             self.model = self.model_id.model
         else:
             self.model = False
-
-    def build_expression(self, field_name, sub_field_name, null_value):
-        """Returns a placeholder expression for use in a template field,
-        based on the values provided in the placeholder assistant.
-
-        :param field_name: main field name
-        :param sub_field_name: sub field name (M2O)
-        :param null_value: default value if the target value is empty
-        :return: final placeholder expression """
-        expression = ''
-        if field_name:
-            expression = "${object." + field_name
-            if sub_field_name:
-                expression += "." + sub_field_name
-            if null_value:
-                expression += " or '''%s'''" % null_value
-            expression += "}"
-        return expression
-
-    @api.onchange('model_object_field', 'sub_model_object_field', 'null_value')
-    def onchange_sub_model_object_value_field(self):
-        if self.model_object_field:
-            if self.model_object_field.ttype in ['many2one', 'one2many', 'many2many']:
-                model = self.env['ir.model']._get(self.model_object_field.relation)
-                if model:
-                    self.sub_object = model.id
-                    self.copyvalue = self.build_expression(self.model_object_field.name, self.sub_model_object_field and self.sub_model_object_field.name or False, self.null_value or False)
-            else:
-                self.sub_object = False
-                self.sub_model_object_field = False
-                self.copyvalue = self.build_expression(self.model_object_field.name, False, self.null_value or False)
-        else:
-            self.sub_object = False
-            self.copyvalue = False
-            self.sub_model_object_field = False
-            self.null_value = False
 
     def unlink(self):
         self.unlink_action()
